+++ Backend/auth.js
@@ -1,0 +1,81 @@
+import express from "express";
+import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
+import cookieParser from "cookie-parser";
+import db from "./db.js"
+import cors from "cors";
+
+const router = express.Router();
+// router.use(express.json());
+// router.use(cookieParser());
+// router.use(cors({ origin: "http://localhost:5173", credentials: true }));
+
+
+const JWT_SECRET = process.env.JWT_SECRET_KEY;
+
+function authenticateTokenCheck(req, res, next) {
+
+    const token = req.cookies.token;
+    if (!token) {
+        return res.status(401).json({ error: "Unauthorized" });
+    }
+
+    try {
+        const decoded = jwt.verify(token, JWT_SECRET);
+        req.userId = decoded.userId;
+        next();
+    } catch {
+        res.status(401).json({ error: "Invalid token" });
+    }
+}
+
+router.post("/register", async (req, res) => {
+    const { email, password } = req.body;
+
+    const hashed = await bcrypt.hash(password, 10);
+
+    try {
+        const result = await db.query(
+            "INSERT INTO users (email, password_hash) VALUES ($1, $2) RETURNING id, email",
+            [email, hashed]
+        );
+
+        res.json(result.rows[0]);
+    } catch (err) {
+        if (err.code === "23505") {
+            return res.status(400).json({ error: "Email already exists" });
+        }
+        res.status(500).json({ error: "Server error" });
+    }
+});
+
+router.post("/login", async (req, res) => {
+    const { email, password } = req.body;
+
+    const result = await db.query("SELECT * FROM users WHERE email = $1", [email]);
+    const user = result.rows[0];
+
+    if (!user) return res.status(400).json({ error: "Invalid email or password" });
+
+    const isMatch = await bcrypt.compare(password, user.password_hash);
+    if (!isMatch) return res.status(400).json({ error: "Invalid email or password" });
+
+
+    const token = jwt.sign({ userId: user.id }, JWT_SECRET, { expiresIn: "1h" });
+
+    res.cookie("token", token, {
+        httpOnly: true,
+        sameSite: "lax",
+        secure: false,
+    });
+
+    res.json({ message: "Logged in" });
+});
+
+router.post("/logout", (req, res) => {
+    res.clearCookie("token");
+    res.json({ message: "Logged out" });
+});
+
+
+export {router as default, authenticateTokenCheck}
