+++ Backend/db.js
@@ -1,0 +1,17 @@
+// db.js
+import pg from "pg";
+import dotenv from "dotenv";
+
+const { Pool } = pg;
+dotenv.config();
+
+const pool = new pg.Pool({
+    user: process.env.PG_USER,
+    host: process.env.PG_HOST,
+    database: process.env.PG_DATABASE,
+    password: process.env.PG_PASSWORD,
+    port: process.env.PG_PORT,
+});
+
+
+export default pool;
